修改限购活动

+++ esp-domain/src/main/java/com/apass/esp/domain/entity/LimitGoodsSku.java
@@ -12,7 +12,7 @@ public class LimitGoodsSku extends CreatedUser{
     private Long limitNum;
     private Long sortNo;
     private String url;
-    private int time;//离活动开始时间或活动开始时间与服务器时间差
+    private Long time;//离活动开始时间或活动开始时间与服务器时间差
     public Long getId() {
         return id;
     }
@@ -73,10 +73,10 @@ public class LimitGoodsSku extends CreatedUser{
     public void setUrl(String url) {
         this.url = url;
     }
-	public int getTime() {
+	public Long getTime() {
 		return time;
 	}
-	public void setTime(int time) {
+	public void setTime(Long time) {
 		this.time = time;
 	}
